临时提交！！！
git-svn-id: http://jrt.choury.com/svn/myos@50 2cdf69c7-79f9-4451-ab89-d57ae126d6d6

+++ kernel/sbrk.c
@@ -8,5 +8,5 @@
  */
 void * sys_sbrk(int incr) {
 //    errno = ENOMEM;
-    return (u32)PROTABLE[curpid].heap+incr;
+    return PROTABLE[curpid].heap+incr;
 }
